v1.0.291 Fix compile error v1.0.290 : ticks int -> float dans tooltip GUI
=== ERREUR REPORTEE (build Alexis) ===

GuiFurnaceNexus.java:513: error: incompatible types: possible lossy conversion
from float to int
    int ticks = tile.getEffectiveMaxCookTime();

Cause : v1.0.290 a change la signature de getEffectiveMaxCookTime() de
int -> float pour supporter le multi-items par tick. Mais cet unique call site
externe dans GuiFurnaceNexus.java ligne 513 utilisait encore 'int ticks ='.

J'avais oublie de le mettre a jour lors du commit v1.0.290. Oubli de review
Superpowers - je n'avais pas grep-e tous les callers de la methode avant de
pousser.

=== FIX ===

GuiFurnaceNexus.java ligne 513-515 : int ticks -> float ticks.
Les calculs aval (ticks / 20.0F et 20.0F / ticks) fonctionnent deja avec
n'importe quel float. Aucun autre changement requis.

Nouveau comportement affichage tooltip :
  Ancien (plafonne) : 'Cuit en: 0.05s (8 SP)' pour Pallanutro+8
  Nouveau (precis)  : 'Cuit en: 0.05s (8 SP)' puis 'Taux: 193 items/s'
  Infinite+8 :       'Cuit en: 0.03s (8 SP)' puis 'Taux: 348 items/s'

=== VALIDATION ===

Grep 'getEffectiveMaxCookTime' : 6 sites au total.
  - GuiFurnaceNexus ligne 515    : float ticks = ... (FIXED)
  - TileFurnaceNexus ligne 271   : declaration methode (OK)
  - TileFurnaceNexus ligne 306   : 200.0F / getEffective... (OK, float)
  - TileFurnaceNexus ligne 457   : maxCookTime = getEffective... (OK, maxCookTime est float)

Brace balance OK :
  GuiFurnaceNexus.java  : 89/89, 461/461
  TileFurnaceNexus.java : 150/150, 692/692

Rebuild doit passer maintenant.

+++ mod-source/src/main/java/com/nexusabsolu/mod/gui/furnaces/GuiFurnaceNexus.java
@@ -510,7 +510,9 @@ public class GuiFurnaceNexus extends GuiContainer {
                 ? tile.getCookProgress() * 100 / tile.getMaxCookTime() : 0;
             lines.add("\u00a7eCuisson\u00a7r: " + pct + "%");
             // Temps de cuisson effectif (avec upgrades)
-            int ticks = tile.getEffectiveMaxCookTime();
+            // v1.0.290 : getEffectiveMaxCookTime() retourne maintenant un float
+            // (peut etre < 1 tick pour les tres hauts tiers avec boosters).
+            float ticks = tile.getEffectiveMaxCookTime();
             float seconds = ticks / 20.0F;
             int speedCount = tile.getSpeedBoosterCount();
             int effCount = tile.getEfficiencyCount();
